Retrieves files from server -> loads into database
This file downloads files from a server via FTP and then uses a PDO (PHP Data Object) to load the data into a database. The data retrieved from the server is loaded into the database as three different tables: tbgps, tbpictures, and tbaudio. SQL is wrapped in the PDO and functions as a DDL to create and manipulate the data.